Save csv log of calculated histograms
Temporary solution: hist_detect_calc saves a csv of histograms it
calculates in each frame. This will be used for determining standard
histograms for later use in detection/comparison. Later a modified
version of this will be implemented as part of the DEBUG mode. Something
like saving the histogram of all matched (detected) contours to a csv.

Other:
 - Fixed bug in hist_detect_calc where it did not draw contours mask
 when not in debug mode.
 - Minor changes to main.cpp