Config for hist-detect; detection mode conditions
Updated config to include minimum histogram difference thresholds
required for histogram detection (in hist_detect). This required
changing the config set and show functions and the config.txt file.

The detection mode to be used (hue or histogram) was also added to the
config file.

Related to the above, various conditions were added to main.cpp to only
execute certain blocks of code based on which detection mode is in use.

Other:
 - sh_detect will terminate early if the car provided has negative hue
   or delta values (which indicate it can only be detected using
   histogram detection)
 - Minor edits to hist_detect (mainly debug printing outputs)