Background subtraction & histogram calibrarion
Main now calculates a difference image by comparing a background image
taken at the start of tracking with each frame. This is used to form the
global mask and the HSV image used for histogram calculation.

Standard histograms have also been calibrated for this process and are
included in the hist_config.txt file.